remove departure indexing code, use linear searches.
Binary search cannot work if we mix frequency and one-off TripTimes together.
It was already disabled because we allow combining trips from different service IDs on the same tripPattern.
Even with real-time updates, it became necessary to keep sorted trip lists at each stop (the indexes) for binary searching.
All the complexity of keeping indexes up to date does not appear to be worth the apparently minor speed improvement.